backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
motor
pymongo
psutil
python-jose[cryptography]
passlib[bcrypt]
python-multipart